pyproject.toml
[project]
name = "cacao-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "ultralytics>=8.3.203",
]
